saving .data PDFs when running from other folder
Previous code gave error when running from hierarchical folder (e.g.
GABA_file = '.\name\MRI\raw.data) I got an error when Gannet tried to
save the pdf for .data file.

Error was: saveas at 115 Invalid or missing path:
./MRSfit_140102/.\7011-0124\MRI\raw_008.data.pdf

So it obviously didn't rewrite the path properly for the pdf here, but
it IS important to get both folder and filename as a lot of the .data
files have similar names (e.g. raw_001.data). This change works for me
for now, might not be most elegant